+++ Core/vadebug.cpp
@@ -27,13 +27,13 @@ namespace vamiga {
 
 #define LOG_FLAG_ENTRY(name, dflt, help) \
     { #name, help, false, \
-      []() -> long { return (long)name; }, \
-      [](long value) { name = value; } },
+      []() -> isize { return (isize)name; }, \
+      [](isize value) { name = value; } },
 
 #define DEBUG_FLAG_ENTRY(type, name, dflt, help) \
     { #name, help, std::is_same_v<type, bool>, \
-      []() -> long { return (long)name; }, \
-      [](long value) { name = (type)value; } },
+      []() -> isize { return (isize)name; }, \
+      [](isize value) { name = (type)value; } },
 
 const std::vector<FlagInfo> logFlags = { VA_LOG_FLAGS(LOG_FLAG_ENTRY) };
 const std::vector<FlagInfo> debugFlags = { VA_DEBUG_FLAGS(DEBUG_FLAG_ENTRY) };
